% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORproject-package.R
\docType{package}
\name{CORproject-package}
\alias{CORproject}
\alias{CORproject-package}
\title{CORproject: Read PNG Using COR And Build A Summary Rmd}
\description{
Provide data and code for optical character recognition of data in png format.
}
\author{
\strong{Maintainer}: Nicolas RODE \email{nicolas.rode@inrae.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Camille IMBERT \email{camille.imbert@inrae.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Diego AYALA \email{diego.ayala@ird.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Stanislas ZANVO \email{zanvostanislas@yahoo.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
