% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpng.R
\name{createpng}
\alias{createpng}
\title{createpng}
\usage{
createpng(
  csvpath = "data/FitnessExperimentalDesign_G60Final.csv",
  outputpng = here::here("data", "testnew.png")
)
}
\arguments{
\item{csvpath}{path to the .csv file}

\item{outputpng}{path to the output png}
}
\value{
A png with the simulated data
}
\description{
Function that takes a png and return a vector with the rows found by tesseract
}
\examples{
createpng(csvpath="data/FitnessExperimentalDesign_G60Final.csv", outputpng="data/test.png")
}
