% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_new_function.R
\name{my_new_function}
\alias{my_new_function}
\title{Takes in a dataset and adds a new column with new values}
\usage{
my_new_function(dataset, new_variable_name, new_variable_value)
}
\arguments{
\item{dataset}{Any input dataset we want to add a column to}

\item{new_variable_name}{The name of the new column ("text string")}

\item{new_variable_value}{The value that the new column should take}
}
\value{
The input dataset with the new column added
}
\description{
Takes in a dataset and adds a new column with new values
}
\examples{
corproject::my_new_function(cars, "hello", "abc")
}
