% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocrpng.R
\name{ocrpng}
\alias{ocrpng}
\title{Title}
\usage{
ocrpng(
  impagepath = "data/test.png",
  n.col = NULL,
  lang = "eng",
  header = TRUE,
  cleaning = FALSE,
  outcsv = "data/fitnessOCR.csv"
)
}
\arguments{
\item{cleaning}{}
}
\value{

}
\description{
Title
}
\examples{

}
