% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_new_function.R
\name{my_new_function}
\alias{my_new_function}
\title{a new function I created}
\usage{
my_new_function(dataset, new_var_name, new_var_value)
}
\arguments{
\item{dataset}{The dataset to add a new column to}

\item{new_var_name}{The name of the new column}

\item{new_var_value}{The value for the new column to take}
}
\value{
The input dataset with a new variable called new_var_name with the values of new_var_value
}
\description{
a new function I created
}
\examples{
new_ds <- my_new_function(cars, "new_variable_name", "test_value")
}
